requirements.txt
asgiref==3.9.1
boto3==1.40.16
botocore==1.40.16
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
django-cors-headers==4.7.0
django-storages==1.14.6
djangorestframework==3.16.1
drf-yasg==1.21.10
futures==3.0.5
goslate==1.5.4
idna==3.10
inflection==0.5.1
jmespath==1.0.1
packaging==25.0
pillow==11.3.0
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
